requirements.txt
pyautogen~=0.2.0b4
cohere
openai==0.28
midi2audio

